requirements.txt
fastapi
uvicorn[standard]
numpy
faiss-cpu
sentence-transformers
PyPDF2
python-docx
gradio
google-generativeai
requests
pydantic